feature: added App Center. renamed package name. added TextView with version number. added App Center badge to README
