+++ README.md
@@ -2,3 +2,5 @@
 temporary repo for android app
 
 CircleCI - [![CircleCI](https://circleci.com/gh/yks72p/md_second_project.svg?style=svg)](https://circleci.com/gh/yks72p/md_second_project)
+App Center - [![Build status](https://build.appcenter.ms/v0.1/apps/cc1ead8d-d71e-45ba-9640-b1a7e536d2d2/branches/master/badge)](https://appcenter.ms)
+
