+++ app/src/androidTest/java/com/example/trafimau/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.secondproject;
+package com.example.trafimau;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
